refactor(service): 拆分经验等级配置缓存逻辑
